+++ src/design/plone/policy/upgrades.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 from copy import deepcopy
-from plone.app.upgrade.utils import installOrReinstallProduct
 from plone import api
+from plone.app.upgrade.utils import installOrReinstallProduct
+from Products.CMFPlone.interfaces import ISelectableConstrainTypes
 
 import logging
 
@@ -83,10 +84,27 @@ def to_1300(context):
     inside the folder "Documenti e Dati".
 
     This method just ADD THE CONTENT TYPE to the current list of types that you
-    can add inside that folder.
+    can add inside that folder only if it's not already there.
+
+    tp#17807
     """
-    pc = api.portal.get_tool(name="portal_catalog")
-    if pc:
-        pass
-    import pdb
-    pdb.set_trace()
+
+    doc_brains = api.content.find(portal_type='Document')
+    doc_e_dati_list = [x for x in doc_brains if x.Title == "Documenti e dati"]
+    if len(doc_e_dati_list) == 1:
+        doc_e_dati = doc_e_dati_list[0].getObject()
+
+        constraints = ISelectableConstrainTypes(doc_e_dati)
+        allowed_types = constraints.getLocallyAllowedTypes()
+        if "Bando" not in allowed_types:
+            allowed_types.append("Bando")
+            constraints.setLocallyAllowedTypes(allowed_types)
+
+            logger.info("Enabled 'Bando' inside 'Ducumenti e dati' folder.")
+        else:
+            logger.info("'Bando' already enabled in 'Ducumenti e dati' folder,"
+                        " not changes needed.")
+    else:
+        logger.warning("More than one Document with title 'Documenti e dati'. "
+                       "Type 'Bando' inside 'Ducumenti e dati' folder not "
+                       "enabled.")